Never re-use the layout result from the fieldset contents wrapper.
On the layout object side, all children of a fieldset are wrapped inside
an anonymous fieldset contents block. We'll detect during layout which
of the legend children (if any) will become the rendered legend. Unlike
all other in-flow children of a fieldset element, whose containing block
is the wrapper, the containing block of the rendered legend is the
fieldset layout object itself (not the wrapper child), so if we change /
remove the rendered legend, the wrapper won't be marked for layout.

Bug: 1119400
Change-Id: I5b898af8c7b7e3687def6a2333d43a8ba6af09b1